+++ NEWS.md
@@ -1,4 +1,8 @@
-# positron.tutorials (development version)
+# vscode.tutorials (development version)
+
+* Renamed package from `positron.tutorials` to `vscode.tutorials`.
+
+* Removed `renv` configuration.
 
 # positron.tutorials 0.2.1